+++ src/main/java/com/exam/license/exam/utils/SolutionMapper.java
@@ -7,9 +7,9 @@ import java.util.List;
 import java.util.Map;
 
 public class SolutionMapper {
-    public static List<UserAnswer> mapSolutionAnswer(Map<Long, String> solutionMap) {
+    public static List<UserAnswer> mapSolutionAnswer(Map<Integer, String> solutionMap) {
         List<UserAnswer> userAnswers = new ArrayList<>();
-        for (Map.Entry<Long, String> entry : solutionMap.entrySet()) {
+        for (Map.Entry<Integer, String> entry : solutionMap.entrySet()) {
             userAnswers.add(new UserAnswer(Math.toIntExact(entry.getKey()), entry.getValue()));
         }
         return userAnswers;
